backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
groq
python-dotenv
